Chore: Refactor the project
