+++ pyproject.toml
@@ -1,8 +1,9 @@
 [tool.poetry]
-name = "quake_arena_log"
+name = "quake-arena-log"
 version = "0.1.0"
-description = "Log parser for quake 3 arena log"
-authors = ["Sophia"]
+description = "Quake 3 arena log parser"
+authors = ["Thiago Simões Gonçalves <thiagosgoncalves7@gmail.com>"]
+license = "MIT"
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -13,7 +14,6 @@ black = "^21.7b0"
 isort = "^5.9.3"
 flake8 = "^3.9.2"
 darker = "^1.3.0"
-
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
